refactor(backend): extract assistant runner raw SQL to service layer
